+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Security
 ### Fixed
+- Fix encoding and decoding of singular tuples and records (optimization)
+- Fix support of templated records
 ### Added
+- Add support for `unit` type
+- Add support for `Set.set` stdlib type
 ### Changed
 ### Deprecated
 ### Removed